Parallel Port PIN driver for Linux 2.6: ppin
This kernel module will register the /dev/ppin (10, 151) device which
controls up to eight PINs through the first parallel port.

Controlling the PINs is as easy as 'echo Num State >/dev/ppin', where
Num is 0 to 7 and State is one of 'on', 'off'.
For example: "echo 3 on >/dev/ppin" switches the 3rd PIN on.

You can read the status of the PINs with 'cat /dev/ppin'.

This program is free software; you can redistribute it and/or modify it
under the terms of the GNU General Public License as published by the
Free Software Foundation; either version 2 of the License, or (at your
option) any later version.

Authors:
Alexander Barton, <alex@barton.de> (for Linux 2.6, 2009)

This work is heavily(!) based on the "devled" driver written by
Konstantinos Natsakis, <cyfex@mail.com> for Linux 2.2/2.4.

Please note that documentation is mostly lacking at the moment --
use the source, luke ;-)